Improve Data_Sync_Entry terminology
