+++ projects/packages/wp-js-data-sync/changelog/boost-update-igp-phase-5
@@ -1,0 +1,4 @@
+Significance: major
+Type: changed
+
+Changed the Data Sync Entry terminology and updated inline documentation
