+++ projects/packages/wp-js-data-sync/src/contracts/interface-data-sync-entry.php
@@ -1,0 +1,52 @@
+<?php
+
+namespace Automattic\Jetpack\WP_JS_Data_Sync\Contracts;
+
+/**
+ * Data Sync Entry Interface:
+ * ==========================
+ * This interface defines what is a Data Sync Entry.
+ *
+ * The "can" method is used to determine if the class implementing this interface supports a specific operation.
+ */
+interface Data_Sync_Entry {
+
+	/**
+	 * Checks if the implementing class supports the specified method.
+	 *
+	 * @param string $method The name of the method to check (get, set, merge, delete).
+	 * @return bool True if the method is supported, false otherwise.
+	 */
+	public function can( $method );
+
+	/**
+	 * Retrieves the current value of the data sync entry.
+	 *
+	 * @return mixed The current value of the data sync entry.
+	 */
+	public function get();
+
+	/**
+	 * Sets the value of the data sync entry.
+	 *
+	 * @param mixed $value The new value to set for the data sync entry.
+	 * @return mixed The updated value of the data sync entry.
+	 */
+	public function set( $value );
+
+	/**
+	 * Merges a partial value with the current value of the data sync entry.
+	 *
+	 * @param mixed $partial_value The partial value to merge with the current value.
+	 * @return mixed The updated value of the data sync entry after merging.
+	 */
+	public function merge( $partial_value );
+
+	/**
+	 * Deletes the data sync entry.
+	 *
+	 * @return mixed The value of the data sync entry after deletion.
+	 */
+	public function delete();
+
+}
